+++ CHANGELOG.rst
@@ -2,6 +2,12 @@
 Changelog
 =========
 
+0.10.0 (2021-11-05)
+-------------------
+
+* Add backgroud refresh token, jwks and revocation list.
+* Add Flask framework support.
+
 0.8.0 (2021-10-22)
 ------------------